requirements.txt
numpy
matplotlib
pillow
mpl_toolkits
scipy
pandas
opencv-python